requirements.txt
appdirs==1.4.3
asgiref==3.2.10
astroid==2.4.2
autopep8==1.5.4
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.3
defusedxml==0.6.0
distlib==0.3.0
Django==3.1
django-allauth==0.42.0
django-cors-headers==3.4.0
django-rest-auth==0.9.5
djangorestframework==3.11.0
filelock==3.0.12
gunicorn==20.0.3
idna==2.10
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
pep8==1.7.1
pipenv==2018.11.26
pycodestyle==2.6.0
pylint==2.5.3
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.10
virtualenv==20.0.19
virtualenv-clone==0.5.4
whitenoise==5.2.0
wrapt==1.12.1
